Fix serialization
* Added one test which test for some properties after from parsed json
* Removed unneeded SerialName annotations
* Removed most Transient annotations in ResultObjects.kt, as the properties should't be ignored
* Removed constructor params for common object to fix serialization
* Apply format

This commit reverts '#fc2f3bc2a0802dda6e6d66059b7813accdea8e7b' because it broke the possibility
to serialize SpotifyUri from Strings, as it changed the json definition to be a object. This didn't
triggered any tests, as no tests are testing for the uri. They just say it's parsed, so it's fine,
but actually there was the TransientUri in each object...