Fix serialization
* Added one test which test for some properties after from parsed json
* Removed unneeded SerialName annotations
* Removed most Transient annotations in ResultObjects.kt, as the properties should't be ignored
* Removed constructor params for common object to fix serialization
* Apply format

This commit reverts '#2c59dc548db5c2ad441355a9aac7a989541e4b1b' because it broke the possibility
to serialize SpotifyUri from Strings, as it changed the json definition to be a object. This didn't
triggered any tests, as no tests are testing for the uri. They just say it's parsed, so it's fine,
but actually there was the TransientUri in each object...